+++ src/interfaces/packagingInterfacesAndType.ts
@@ -6,7 +6,7 @@
  */
 
 import { Duration } from '@salesforce/kit';
-import { Connection, SfProject } from '@salesforce/core';
+import { Connection, NamedPackageDir, SfProject } from '@salesforce/core';
 import { QueryResult, SaveResult } from 'jsforce';
 import { Attributes } from 'graphology-types';
 import { PackageProfileApi } from '../package/packageProfileApi';
@@ -165,6 +165,34 @@ export type PackageCreateOptions = {
   path: string;
 };
 
+export type PackageDescriptorJson = Partial<NamedPackageDir> &
+  Partial<{
+    id: string;
+    features: string[];
+    orgPreferences: string[];
+    snapshot: string;
+    unpackagedMetadata: NamedPackageDir;
+    apexTestAccess: { permissionSets: string[] | string; permissionSetLicenses: string[] | string };
+    permissionSetNames: string[];
+    permissionSetLicenseDeveloperNames: string[];
+    branch: string;
+    subscriberPackageVersionId: string;
+    packageId: string;
+    versionName: string;
+  }>;
+
+export type PackageVersionCreateRequest = {
+  Package2Id: string;
+  VersionInfo: string;
+  Tag: string;
+  Branch: string;
+  InstallKey: string;
+  Instance: string;
+  SourceOrg: string;
+  CalculateCodeCoverage: boolean;
+  SkipValidation: boolean;
+};
+
 export type PackageVersionQueryOptions = {
   project: SfProject;
   orderBy: string;
@@ -224,8 +252,7 @@ export type PackageVersionCreateOptions = Partial<
     definitionfile: string;
     installationkey: string;
     installationkeybypass: boolean;
-    package: string;
-    path: string;
+    packageId: string;
     postinstallscript: string;
     postinstallurl: string;
     preserve: boolean;
@@ -239,8 +266,6 @@ export type PackageVersionCreateOptions = Partial<
     versiondescription: string;
     versionname: string;
     versionnumber: string;
-    wait: Duration;
-    pollInterval: Duration;
     profileApi: PackageProfileApi;
   }
 >;