+++ src/package/package.ts
@@ -8,12 +8,14 @@ import { Messages, sfdc } from '@salesforce/core';
 import { AsyncCreatable, Duration } from '@salesforce/kit';
 import { QueryResult } from 'jsforce';
 import { Optional } from '@salesforce/ts-types';
-import { IPackage, PackageOptions, PackagingSObjects } from '../interfaces';
 import {
+  IPackage,
+  PackageOptions,
+  PackagingSObjects,
   PackageInstallOptions,
   PackageInstallCreateRequest,
   PackageIdType,
-} from '../interfaces/packagingInterfacesAndType';
+} from '../interfaces';
 import { listPackages } from './packageList';
 import { getExternalSites, getStatus, installPackage, waitForPublish } from './packageInstall';